Relocate convertMovieResponseToCard function
