+++ app/home/page.tsx
@@ -1,7 +1,7 @@
 import { fetchPopularMovies } from "../lib/http/movie/movie.service";
 
 import { CardList, CardListProps } from "./component/CardList/CardList";
-import { convertMovieResponseToCard } from "./component/utils";
+import { convertMovieResponseToCard } from "./component/Card/utils";
 
 export default async function Home() {
   const movies = await fetchPopularMovies({ page: String(1) });
